Property-test splice loop and instrument its boundaries
Address two PR review warnings.

Testing: factor the accumulation loop into `accumulate_splices` with its
splice and drain side effects injected, then property-test the
`Ok(0)` / `Ok(n)` / broken-pipe-drain / fatal transitions over generated
outcome sequences. This complements the `splice_once_with` EINTR proptest.
Exhaustive Kani model-checking of these paths stays tracked in #87 and #88.

Observability: emit bounded, low-cardinality `tracing` at the three splice
boundaries — a `debug` event at the support-detection fallback, a `trace`
event per `EINTR` retry, and a `debug` event carrying `bytes_transferred` at
the broken-pipe drain. Following the library convention the crate emits
instrumentation but installs no subscriber; the Python command boundary owns
subscriber configuration and higher-level command observation. Add the
`tracing` dependency and revise the splice-loop/drain contract accordingly,
superseding the earlier "no telemetry surface" note.

Move `splice.rs` to `splice/mod.rs` and extract the test module into
`splice/tests.rs` so both files stay within the 400-line module limit.
