+++ rust/cuprum-rust/Cargo.toml
@@ -13,6 +13,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 pyo3 = { version = "0.29.0", features = ["extension-module"] }
 thiserror = "2.0.18"
+tracing = { version = "0.1.44", default-features = false, features = ["std"] }
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2.186"
